backend/requirements.txt
# Backend Requirements for Deepfake Detection System

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML - PyTorch
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0

# Training & Visualization
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# WebSocket for real-time video
websockets>=12.0

# CORS
python-dotenv>=1.0.0

# Utilities
aiofiles>=23.2.0
